requirements.txt
streamlit
moviepy
google-cloud-speech
google-cloud-texttospeech
openai
ffmpeg-python